moved serialization into separate file
