+++ packages/wasm/tests/test_compilation.rs
@@ -32,23 +32,3 @@ fn wasm_module_from_bytecode_compile() {
 
     assert!(result.is_ok());
 }
-
-#[test]
-fn wasm_module_serialization() {
-    let test_path = get_tests_path().unwrap();
-    let path = test_path.into_os_string().into_string().unwrap();
-
-    let module_path = format!("{path}/subinvoke/00-subinvoke/implementations/as/wrap.wasm");
-
-    let module_bytes = fs::read(Path::new(&module_path)).unwrap();
-
-    let module = WasmModule::WasmBytecode(module_bytes);
-
-    let module = module.compile().unwrap();
-
-    let result = module.serialize().unwrap();
-
-    let result = result.deserialize();
-
-    assert!(result.is_ok());
-}
